pyproject.toml
[tool.poetry]
name = "chest-x-ray-ddpm"
version = "0.1.0"
description = ""
authors = ["Hippolyte Mayard"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.0"
torch = {url = "https://download.pytorch.org/whl/cu116/torch-1.13.1%2Bcu116-cp310-cp310-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu116/torchvision-0.14.1%2Bcu116-cp310-cp310-linux_x86_64.whl"}
einops = "^0.6.1"
pillow = "^9.5.0"
tqdm = "^4.65.0"
jupyter = "^1.0.0"
matplotlib = "^3.7.2"


[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
line_length = 120
profile = 'black'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
